fix useEffect triggering error several times
